Add initial API for Gothenburg
This commit adds a happy path login flow and a rungbg script for
doing the login and then fetching the logged in person's children.

The login flow is based on the network interactions of a browser when
logging in. There are a few... quirks... in the login to Hjärntorget in
particular a few responses from the server ends with a html response
containg a very simple form and a script that post the form on load.

This is likely because there are at least 3 different domains invovled
when logging in to Hjarntorget:
 - hjarntorget.gothenburg.se
 - auth.gothenburg.se
 - m00-mg-local.idp.funktionstjanster.se

The data is passed between the user's browser and the different sites
via the mentioned form posting mechanism.

Anyway run `node ./rungbg.js <pnr>` if you know who Gothenburg thinks
your kids are :)